requirements.txt
python-telegram-bot[ext]~=20.3
asyncpg>=0.27.0
python-dotenv>=0.19.0
apscheduler>=3.10.0
nest_asyncio>=1.5.6
aiofiles>=23.1.0
uvloop>=0.17.0